+++ CHANGELOG.md
@@ -18,4 +18,5 @@
 
 ## Version 2.0
 
+* Implemented Exercise 16 - Merging Streams.
 * Implemented Exercise 18 - Aggregating Flink Data using Windowing.